refactor(cli): move parser types to parser folder
