+++ src/cli/parser/types.ts
@@ -1,4 +1,13 @@
-import type { OptionSpecType, SpecSchema } from '@/cli/types';
+export type OptionSpecType = 'boolean' | 'string' | 'stringArray';
+
+export interface OptionSpec {
+  description: string;
+  longFlag: string;
+  shortFlag?: string;
+  type?: OptionSpecType;
+}
+
+export type SpecSchema = Record<string, OptionSpec>;
 
 type OptionSpecTypeMap = Record<OptionSpecType, unknown> & {
   boolean: boolean;
